Rework consumer callback to take a producer object
Big change is split Kafka wrapping API into producer and consumer.
Producer half is constructed at the beginning of the builder, and is now
passed to consumer callback functions - they frequently send messages,
but the function is written before the Kafka object is constructed.

Also have a semi placeholder interface for the consumer half, which
currently only allows closing the connection.  In the future, there
might be more API calls, and having the inferface now should make this
easier to implement.